Adiciona DTOs e repositório para gerenciamento de empréstimos

+++ Services/Livro/LivroService.cs
@@ -1,6 +1,6 @@
 ﻿using Sistema_Biblioteca.Controllers.Dtos;
 using Sistema_Biblioteca.Entities;
-using Sistema_Biblioteca.Repositories.Interface;
+using Sistema_Biblioteca.Repositories.Livros;
 using Sistema_Biblioteca.Services.Interface;
 
 namespace Sistema_Biblioteca.Services
@@ -14,40 +14,43 @@ namespace Sistema_Biblioteca.Services
             this.livroRepository = livroRepository;
         }
 
-        public async Task<ListarLivrosDto> CreateLivroAsync(CadastrarLivroDto livro)
+        public async Task<LivroReponseDto> CreateLivroAsync(CadastrarLivroDto livro)
         {
             var request = new Livro { Nome = livro.Nome, DataCadastro = DateTime.UtcNow };
-            await livroRepository.Create(request);
+            await livroRepository.Add(request);
 
-            return new ListarLivrosDto { Id = request.Id, Nome = request.Nome };
+            return new LivroReponseDto { Id = request.Id, Nome = request.Nome };
         }
 
         public async Task DeleteLivroAsync(int id)
         {
+            if (id <= 0)
+                throw new ArgumentException("O id do livro deve ser maior que zero.");
+
             var livro = await livroRepository.GetById(id) ?? throw new Exception("Livro não encontrado");
             await livroRepository.Delete(livro);
         }
 
-        public async Task<IEnumerable<ListarLivrosDto>> GetAllLivrosAsync()
+        public async Task<IEnumerable<LivroReponseDto>> GetAllLivrosAsync()
         {
             var livros = await livroRepository.GetAll();
 
-            return livros.Select(livro => new ListarLivrosDto { Id = livro.Id, Nome = livro.Nome, DataCadastro = livro.DataCadastro }).ToList();
+            return livros.Select(livro => new LivroReponseDto { Id = livro.Id, Nome = livro.Nome, DataCadastro = livro.DataCadastro }).ToList();
         }
 
-        public async Task<ListarLivrosDto> GetLivroByIdAsync(int id)
+        public async Task<LivroReponseDto> GetLivroByIdAsync(int id)
         {
             var livro = await livroRepository.GetById(id) ?? throw new Exception("Livro não encontrado");
-            return new ListarLivrosDto { Id = livro.Id, Nome = livro.Nome, DataCadastro = livro.DataCadastro };
+            return new LivroReponseDto { Id = livro.Id, Nome = livro.Nome, DataCadastro = livro.DataCadastro };
         }
 
-        public async Task<ListarLivrosDto> UpdateLivroAsync(int id, AtualizarLivroDto dto)
+        public async Task<LivroReponseDto> UpdateLivroAsync(int id, AtualizarLivroDto dto)
         {
             var livro = await livroRepository.GetById(id) ?? throw new Exception("Livro não encontrado");
             livro.Nome = dto.Nome;
-            
+
             await livroRepository.Update(livro);
-            return new ListarLivrosDto { Id = livro.Id, Nome = livro.Nome };
+            return new LivroReponseDto { Id = livro.Id, Nome = livro.Nome };
         }
     }
 }
